fix(db): change Usuario propierty telefono
Se Cambio el propiedad de INT a BIGING dado que no permitia valores de
numero como 3213424321

+++ init.sql
@@ -3,7 +3,7 @@ CREATE TABLE Usuario (
   nombre varchar(255) NOT NULL,
   apellido varchar(255) NOT NULL,
   email varchar(255) UNIQUE NOT NULL,
-  telefono integer NOT NULL,
+  telefono BIGINT NOT NULL,
   password varchar(255) NOT NULL
 );
